backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-sqlalchemy==3.0.3
flask-login==0.6.2
python-dotenv==1.0.0
openai==0.27.8
Werkzeug==2.2.3
requests==2.28.2
gunicorn==20.1.0